# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Planificacion Horaria",
        "description": "Gestion de horarios de clases",
        "contact": {
            "email": "contacto@tudominio.com"
        },
        "license": {
            "name": "Licencia",
            "url": "https://www.terciarioUrquiza.com.ar/planificacionHoraria"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/horarios": {
            "get": {
                "tags": [
                    "Horario"
                ],
                "summary": "Obtener todos los horarios",
                "description": "Devuelve todos los horarios",
                "operationId": "getHorarios",
                "responses": {
                    "200": {
                        "description": "Horarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Horario"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los horarios"
                    }
                }
            }
        },
        "/api/horarios/{id}": {
            "get": {
                "tags": [
                    "Horario"
                ],
                "summary": "Obtener horario por id",
                "description": "Devuelve un horario",
                "operationId": "getHorarioPorId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Horario"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe el horario"
                    },
                    "500": {
                        "description": "Error al obtener el horario"
                    }
                }
            }
        },
        "/api/horarios/guardar": {
            "post": {
                "tags": [
                    "Horario"
                ],
                "summary": "Guardar horario",
                "description": "Guardar un horario",
                "operationId": "guardarHorario",
                "requestBody": {
                    "description": "Horario a guardar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioData"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Horario creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HorarioData"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al guardar el horario"
                    }
                }
            }
        },
        "/api/horarios/actualizar/{id}": {
            "put": {
                "tags": [
                    "Horario"
                ],
                "summary": "Actualizar horario",
                "description": "Actualizar un horario",
                "operationId": "actualizarHorario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Horario a actualizar",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/HorarioData"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Horario actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HorarioData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe el horario"
                    },
                    "500": {
                        "description": "Error al actualizar el horario"
                    }
                }
            }
        },
        "/api/horarios/eliminar/{id}": {
            "delete": {
                "tags": [
                    "Horario"
                ],
                "summary": "Eliminar horario",
                "description": "Eliminar un horario",
                "operationId": "eliminarHorario",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id del horario",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Horario eliminado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HorarioData"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe el horario"
                    },
                    "500": {
                        "description": "Error al eliminar el horario"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HorarioData": {
                "title": "HorarioData",
                "description": "Esquema del objeto HorarioData",
                "properties": {
                    "inicio": {
                        "description": "Hora de inicio de la clase",
                        "type": "time",
                        "example": "19:20:00"
                    },
                    "fin": {
                        "description": "Hora de fin de la clase",
                        "type": "time",
                        "example": "20:40:00"
                    },
                    "dia": {
                        "description": "Dia de la clase",
                        "type": "enum",
                        "enum": [
                            "lunes",
                            "martes",
                            "miercoles",
                            "jueves",
                            "viernes"
                        ]
                    }
                },
                "type": "object"
            },
...
            "Horario": {
                "title": "Horario",
                "description": "Esquema del objeto Horario",
                "properties": {
                    "inicio": {
                        "description": "Hora de inicio de la clase",
                        "type": "time"
                    },
                    "fin": {
                        "description": "Hora de fin de la clase",
                        "type": "time"
                    },
                    "dia": {
                        "description": "Dia de la clase",
                        "type": "enum",
                        "enum": [
                            "lunes",
                            "martes",
                            "miercoles",
                            "jueves",
                            "viernes"
                        ]
                    },
                    "fecha_creacion": {
                        "description": "Fecha de creacion del horario",
                        "type": "string",
                        "format": "date-time"
                    },
                    "fecha_modificacion": {
                        "description": "Fecha de modificacion del horario",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    }
}